fix(6): c

+++ 6/C.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-  int area;
+  double area;
   cin >> area;
   cout << fixed << setprecision(10) << sqrt(area / M_PI) * 2 << endl;
 }
